accounts/authentication.py
from accounts.models import User, Token


class PasswordlessAuthenticationBackend(object):

    def authenticate(self, request, uid):
        try:
            token = Token.objects.get(uid=uid)
            return User.objects.get(email=token.email, identifier=token.identifier)
        except Token.DoesNotExist:
            return None
        except User.DoesNotExist:
            return User.objects.create(email=token.email, identifier=token.identifier)

    def get_user(self, email):
        try:
            return User.objects.get(email=email)
        except User.DoesNotExist:
            return None
